move some functions into qusb mod
